+++ programs/stake-vault/src/state/stake_config.rs
@@ -8,8 +8,8 @@ pub struct StakeConfig {
     pub deposit_mint: Pubkey,
     pub reward_mint: Pubkey,
     pub base_yield_rate: u64, 
-    #[max_len(5, 5)]
-    pub lock_durations: Vec<u32>, // Lock days
+    #[max_len(5)]
+    pub lock_durations: Vec<u64>, // Lock days
     pub total_staked: u64,
     pub total_positions: u64,
     pub bump: u8,
